# docs/swagger.yaml
basePath: /v1
definitions:
  api.ErrorDetail:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorDetail'
      metadata:
        $ref: '#/definitions/api.ResponseMetadata'
      success:
        type: boolean
    type: object
...
  api.PaginatedResponse:
    properties:
      data: {}
      metadata:
        $ref: '#/definitions/api.ResponseMetadata'
      pagination:
        $ref: '#/definitions/api.PaginationMetadata'
      success:
        type: boolean
    type: object
  api.PaginationMetadata:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      next_offset:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  api.ResponseMetadata:
    properties:
      timestamp:
        type: string
      trace_id:
        type: string
    type: object
  api.SuccessResponse:
    properties:
      data: {}
      metadata:
        $ref: '#/definitions/api.ResponseMetadata'
      success:
        type: boolean
    type: object
...
  services.CreateTradingRequest:
    properties:
      api_key:
        example: your_api_key_here
        minLength: 1
        type: string
      api_secret:
        example: your_api_secret_here
        minLength: 1
        type: string
      name:
        example: My Trading Account
        maxLength: 100
        minLength: 1
        type: string
      type:
        example: binance
        type: string
    required:
    - api_key
    - api_secret
    - name
    - type
    type: object
...
  services.TradingResponse:
    properties:
      api_key:
        description: Masked in production
        type: string
      created_at:
        type: string
      id:
        type: string
      info:
        additionalProperties: true
        type: object
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  services.UpdateTradingRequest:
    properties:
      api_key:
        example: updated_api_key_12345
        minLength: 1
        type: string
      api_secret:
        example: updated_api_secret_67890
        minLength: 1
        type: string
      name:
        example: My Updated Trading Account
        maxLength: 100
        minLength: 1
        type: string
      status:
        enum:
        - active
        - inactive
        example: active
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@tiris.ai
    name: Tiris API Support
    url: https://tiris.ai/support
  description: A Go-based microservice for quantitative trading data management, providing
    RESTful APIs for user management, trading integration, and trading operations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://tiris.ai/terms
  title: Tiris Backend API
  version: "1.0"
paths:
...
  /admin/tradings:
    get:
      description: Lists all trading configurations with pagination (admin
        only)
      parameters:
      - default: 100
        description: Number of tradings to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of tradings to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all tradings
      tags:
      - Tradings
  /admin/tradings/{id}:
    get:
      description: Retrieves a trading by ID (admin only)
      parameters:
      - description: Trading ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TradingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get trading by ID
      tags:
      - Tradings
...
  /tradings:
    get:
      description: Retrieves all trading configurations for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.TradingResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user tradings
      tags:
      - Tradings
    post:
      consumes:
      - application/json
      description: Creates a new trading configuration for the authenticated
        user
      parameters:
      - description: Create trading request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateTradingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.TradingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new trading
      tags:
      - Tradings
  /tradings/{id}:
    delete:
      description: Deletes a trading configuration (must belong to authenticated
        user)
      parameters:
      - description: Trading ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete trading
      tags:
      - Tradings
    get:
      description: Retrieves a specific trading configuration by ID (must
        belong to authenticated user)
      parameters:
      - description: Trading ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TradingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get trading by ID
      tags:
      - Tradings
    put:
      consumes:
      - application/json
      description: Updates an existing trading configuration (must belong
        to authenticated user)
      parameters:
      - description: Trading ID
        in: path
        name: id
        required: true
        type: string
      - description: Update trading request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateTradingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TradingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update trading
      tags:
      - Tradings
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT Bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...